tools/windows_vision_bridge_requirements.txt
numpy
opencv-python
ultralytics
PyYAML
